requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain-community==0.0.13
langchain-deepseek==0.1.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
